Reorganised the internal structure.
Replaced the base.py and utils.py files with constants.py, utils.py
and validation.py so it is easier to see where everything can be
found.

Added a parameter mapping table to deal with maxResults having three
different default values depending on whether the records being
returned contain observations, checklists or observers.

Split the mixin used for testing maxResults intro three, again
depending whether the test is for observations, checklists or
observers. In particular this means the overflow tests - when the
value is higher than the maximum allowed in the API - can be more
precise.