Removed the section on locales and added one on formats.
Removed the section listing all the locales that are supported. The
list has more to do with eBird rather than the package and the list
is easily found in the files or docstrings.

Added a section reporting that although some eBird API end-point
allow data to be downloaded in CSV format the package only supports
JSON - it's easy enough to convert to CSV if needed.