Refactored validation tests.
Split test_validation into a separate suite for each function in the
validation module. This makes it easier to see what's missing.

The tests are white-box tests. Most of the function are based on
clean_code() and clean_codes() so the tests for these functions were
not duplicated in the functions that called them.

This closes #5.