Added validation for locations.
The eBird API documentation states that when fetching observations a
list of up to 10 locations may be used. What is not mentioned is that
a list of up to 10 regions or countries may be used also. However this
is not encouraged to limit the load on the server.

The clean_locations function and associated tests were added in case
this loophole is closed and only locations may be used in lists.