Setup tests no longer run twice.
The unit tests were factored out into a separate module in an earlier change.
That means the test_suite function used to discover the tests could be
removed as the module path, 'tests.unit' could be passed to the test_suite
parameter in the setup() function. This also solved the problem of the
tests being run twice (the reason why was not tracked down).