Query parameter, checklistSort is replaced by rankedBy, in get_top_100().
As part of the process of writing the updates for the official documentation
of the eBird API the checklistSort query paramater for GET Top 100 end-point
was found not to be working. This was replaced by a new query parameters,
rankedBy. This commit updated the get_top_100() function to support the
new parameter.

This closes #17.